# example-project/cmd/web-api/docs/swagger.yaml
basePath: '{{.BasePath}}'
definitions:
  account.AccountResponse:
    properties:
      address1:
        example: 221 Tatitlek Ave
        type: string
      address2:
        example: 'Box #1832'
        type: string
      archived_at:
        $ref: '#/definitions/web.TimeResponse'
        type: object
      billing_user_id:
        example: d69bdef7-173f-4d29-b52c-3edc60baf6a2
        type: string
      city:
        example: Valdez
        type: string
      country:
        example: USA
        type: string
      created_at:
        $ref: '#/definitions/web.TimeResponse'
        type: object
      id:
        example: c4653bf9-5978-48b7-89c5-95704aebb7e2
        type: string
      name:
        example: Company Name
        type: string
      region:
        example: AK
        type: string
      signup_user_id:
        example: d69bdef7-173f-4d29-b52c-3edc60baf6a2
        type: string
      status:
        $ref: '#/definitions/web.EnumResponse'
        type: object
      timezone:
        example: America/Anchorage
        type: string
      updated_at:
        $ref: '#/definitions/web.TimeResponse'
        type: object
      zipcode:
        example: "99686"
        type: string
    type: object
  account.AccountUpdateRequest:
    properties:
      address1:
        example: 221 Tatitlek Ave
        type: string
      address2:
        example: 'Box #1832'
        type: string
      billing_user_id:
        example: d69bdef7-173f-4d29-b52c-3edc60baf6a2
        type: string
      city:
        example: Valdez
        type: string
      country:
        example: USA
        type: string
      id:
        example: c4653bf9-5978-48b7-89c5-95704aebb7e2
        type: string
      name:
        example: Company Name
        type: string
      region:
        example: AK
        type: string
      signup_user_id:
        example: d69bdef7-173f-4d29-b52c-3edc60baf6a2
        type: string
      status:
        enum:
        - active
        - pending
        - disabled
        example: disabled
        type: string
      timezone:
        example: America/Anchorage
        type: string
      zipcode:
        example: "99686"
        type: string
    required:
    - id
    type: object
...
  web.EnumOption:
    properties:
      selected:
        example: true
        type: boolean
      title:
        example: Active Etc
        type: string
      value:
        example: active_etc
        type: string
    type: object
  web.EnumResponse:
    properties:
      options:
        items:
          $ref: '#/definitions/web.EnumOption'
        type: array
      title:
        example: Active Etc
        type: string
      value:
        example: active_etc
        type: string
    type: object
...
  web.ErrorResponse:
    properties:
      error:
        type: string
      fields:
        items:
          $ref: '#/definitions/web.FieldError'
        type: array
    type: object
  web.FieldError:
    properties:
      error:
        type: string
      field:
        type: string
    type: object
  web.TimeResponse:
    properties:
      date:
        example: "2019-06-25"
        type: string
      kitchen:
        example: 3:00AM
        type: string
      local:
        example: Tue Jun 25 3:00AM
        type: string
      local_date:
        example: Tue Jun 25
        type: string
      now_rel_time:
        example: 15 hours from now
        type: string
      now_time:
        example: 5 hours ago
        type: string
      rfc1123:
        example: Tue, 25 Jun 2019 03:00:53 AKDT
        type: string
      time:
        example: "03:00:53"
        type: string
      timezone:
        example: America/Anchorage
        type: string
      value:
        example: "2019-06-25T03:00:53.284-08:00"
        type: string
      value_utc:
        example: "2019-06-25T11:00:53.284Z"
        type: string
    type: object
host: '{{.Host}}'
info:
  contact:
    email: support@geeksinthewoods.com
    name: API Support
    url: http://example.com/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://example.com/terms
  title: SaaS Example API
  version: '{{.Version}}'
paths:
  /accounts:
    patch:
      consumes:
      - application/json
      description: Update updates the specified account in the system.
      parameters:
      - description: Update fields
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/account.AccountUpdateRequest'
          type: object
      produces:
      - application/json
      responses:
        "201": {}
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
      security:
      - OAuth2Password: []
      summary: Update account by ID
      tags:
      - account
  /accounts/{id}:
    get:
      consumes:
      - application/json
      description: Read returns the specified account from the system.
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/account.AccountResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
            type: object
      security:
      - OAuth2Password: []
      summary: Get account by ID
      tags:
      - account
...
securityDefinitions:
  BasicAuth:
    type: basic
  OAuth2Password:
    flow: password
    scopes:
      admin: ' Grants administrative privileges with role of admin.'
      user: ' Grants basic privileges with role of user.'
    tokenUrl: /v1/oauth/token
    type: oauth2
swagger: "2.0"